Seed on request instead of on start
