+++ app/router.ts
@@ -6,6 +6,8 @@ export default (app: Application) => {
   router.get('/', controller.public.index)
   router.get('/csrf', controller.public.csrf)
 
+  router.post('/seeder/seed', controller.seeder.seedDatabase)
+
   router.post('/appointment/create', controller.appointment.createAppointment)
   router.get('/appointment/find', controller.appointment.findAppointments)
   router.post('/appointment/find', controller.appointment.findAppointments)
